fix: myBoard 요청에 env경로가 제대로 들어가지 않는 에러 수정
