+++ src/App.tsx
@@ -1,5 +1,6 @@
 import './index.scss';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import styled from 'styled-components/macro';
 
 import Navbar from './components/navbar';
 import Login from './pages/Login/index';
@@ -11,7 +12,7 @@ import MyPage from 'components/myPage/MyPage';
 import SearchForm from 'components/navbar/SearchForm';
 import SignUp from 'pages/SignUp';
 import { useAppSelector } from 'store/hooks';
-import MainWrapper from 'styles/app.styled';
+import devices from 'styles/device';
 
 function App() {
   const loading = useAppSelector((state) => state.loading);
@@ -34,4 +35,40 @@ function App() {
     </BrowserRouter>
   );
 }
+
+const MainWrapper = styled.div.attrs({ className: 'main-wrapper' })`
+  width: 95%;
+  margin: auto;
+
+  @media ${devices.mobileM} {
+    width: 95%;
+    .searchBar {
+      width: 100%;
+    }
+    .card-text {
+      bottom: 3.5vw;
+    }
+    .card-underInfo {
+      bottom: 0.5vw;
+    }
+  }
+
+  @media ${devices.tablet} {
+    width: 70%;
+    .searchBar {
+      width: 110%;
+    }
+    .card-text {
+      bottom: 2.4vw;
+    }
+  }
+
+  @media ${devices.laptopL} {
+    width: 55%;
+  }
+  .card-text {
+    /* bottom: 3.5vw; */
+  }
+`;
+
 export default App;
