+++ src/styles/device.ts
@@ -2,9 +2,9 @@ const sizes = {
   mobileS: '320px',
   mobileM: '375px',
   mobileL: '425px',
-  tablet: '834px',
+  tablet: '832px',
   laptop: '1024px',
-  laptopL: '1440px',
+  laptopL: '1920px',
   desktop: '2560px',
 };
 const devices = {
